Add a base class for chooser factories
